backend/pyproject.toml
[project]
name = "legaltech-backend"
version = "0.1.0"
description = "FastAPI backend for the LegalTech MVP"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.115,<1.0",
  "pymupdf>=1.26,<2.0",
  "python-multipart>=0.0.20,<1.0",
  "uvicorn[standard]>=0.34,<1.0",
  "pydantic-settings>=2.8,<3.0",
  "sqlalchemy>=2.0,<3.0",
]

[project.optional-dependencies]
dev = [
  "alembic>=1.16,<2.0",
  "httpx>=0.28,<1.0",
  "pytest>=8.3,<9.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = ["-p", "no:cacheprovider"]
